all: add UI for picking between tails/scales animals
This adds a UI for picking between tails and scales animals -
it grabs images from Wikipedia and tries to find a reasonably
small one, but no guarantees. It's not the most performant and
instead of using the actual tails/scales list from OSS we should
probably whittle down a smaller list of known safe ones, because
there are a bunch of silly joke ones in there like hamlet or
hippocampus and such. It's fine for a first pass.

The buttons don't do anything because I'm not sure how we
want to actually send thigns to Prometheus yet - they'll probably
end up sending POSTs to the backend, at which point I'll need
some sort of CSRF protection, boo.